Update getText behaviour
